Move user minification to users/common
